application/requirements.txt
ultralytics==8.0.0
opencv-python==4.8.0.74
Pillow==10.0.0
torch==2.0.0
torchvision==0.15.0
ttkbootstrap==1.10.1